Move Stub into it's own file.
